+++ lib/json_cache.dart
@@ -9,4 +9,5 @@ export 'src/json_cache_hollow.dart';
 export 'src/json_cache_local_storage.dart';
 export 'src/json_cache_mem.dart';
 export 'src/json_cache_prefs.dart';
+export 'src/json_cache_sec_storage.dart';
 export 'src/json_cache_wrap.dart';